+++ src/components/FlowPanel/index.tsx
@@ -1,24 +1,67 @@
 import { orchEdgeTypes } from "@/edges";
 import { orchNodeTypes } from "@/nodes";
-import { useMemo } from "react";
+import { useCallback, useMemo, useRef } from "react";
 import ReactFlow, {
   Background,
   BackgroundVariant,
+  Node,
   ReactFlowProps,
+  addEdge,
+  useEdgesState,
+  useNodesState,
+  useReactFlow,
 } from "reactflow";
 import "reactflow/dist/style.css";
+import { IFlowPanel } from "./model";
 
-export const FlowPanel = ({
-  ...rest
-}: Omit<ReactFlowProps, "nodeTypes" | "edgeTypes">) => {
+export const FlowPanel = ({ onDropNewNode, ...rest }: IFlowPanel) => {
   const nodeTypes = useMemo(() => orchNodeTypes, []);
   const edgeTypes = useMemo(() => orchEdgeTypes, []);
 
+  const { screenToFlowPosition } = useReactFlow();
+
+  const onDragOver = useCallback((event: any) => {
+    event.preventDefault();
+    event.dataTransfer.dropEffect = "move";
+  }, []);
+
+  const onDrop = useCallback(
+    (event: any) => {
+      event.preventDefault();
+
+      const type = event.dataTransfer.getData("application/reactflow");
+
+      if (typeof type === "undefined" || !type) {
+        return;
+      }
+
+      const position = screenToFlowPosition({
+        x: event.clientX,
+        y: event.clientY,
+      });
+
+      const newNode: Node = {
+        id: crypto.randomUUID(),
+        type,
+        position,
+        data: {},
+      };
+
+      onDropNewNode?.(newNode);
+    },
+    [screenToFlowPosition]
+  );
+
   return (
     <>
-      <ReactFlow fitView nodeTypes={nodeTypes} edgeTypes={edgeTypes} {...rest}>
-        {/* <Controls />
-        <MiniMap /> */}
+      <ReactFlow
+        {...rest}
+        onDrop={onDrop}
+        onDragOver={onDragOver}
+        nodeTypes={nodeTypes}
+        edgeTypes={edgeTypes}
+        fitView
+      >
         <Background
           color="#292929"
           variant={BackgroundVariant.Dots}