+++ src/components/FlowPanel/model.ts
@@ -1,0 +1,6 @@
+import { Node, ReactFlowProps } from "reactflow";
+
+export interface IFlowPanel
+  extends Omit<ReactFlowProps, "nodeTypes" | "edgeTypes"> {
+  onDropNewNode?: (node: Node) => void;
+}